исправил скачивание картинок

+++ main.py
@@ -74,8 +74,9 @@ def download_txt(url, book_id, filename, folder='books/'):
 def download_image(file_name, url, folder='images/'):
     response = requests.get(url, verify=False)
     response.raise_for_status()
-    result_file_name = f'{file_name}{os.path.splitext(unquote(urlparse(url).path))[1]}'
-    file_path = os.path.join(folder, result_file_name)
+    file_path = os.path.join(folder, os.path.basename(unquote(urlparse(url).path)))
+    if os.path.exists(file_path):
+        return file_path
     with open(file_path, 'wb') as file:
         file.write(response.content)
     logger.info(f'download file: {file_path}')
